Add complex bundles on mocked_plans
- order apps by charm and application name to have always the same
  sequence
